TILA-2287 gql changes (#530)
* TILA-2287 gql changes

* TILA-2287 cleanup

* TILA-2287 gql type change; convert input decimals to strings

* TILA-2287 convert input types

* TILA-2287 remove unnecessary type conversion

* TILA-2287 cleanup

+++ admin-ui/src/component/Unit/queries.ts
@@ -5,7 +5,7 @@ export const UNITS_QUERY = gql`
     $first: Int
     $offset: Int
     $orderBy: String
-    $serviceSector: Float
+    $serviceSector: Decimal
     $nameFi: String
   ) {
     units(
